fix: init PostHog in a client provider, not instrumentation-client (events now send)
Root cause of the 0-events-in-PostHog problem: posthog.init() lived in
instrumentation-client.ts, which Next builds as a separate entry. That gave it
a different posthog-js module instance than the app bundle, so init ran (the
remote config.js loaded — making it look wired) while every capture() in
PostHogPageView / track() hit a second, uninitialized instance and silently
sent nothing. Web analytics stayed at 0 despite real traffic.

Fix mirrors the working kyma setup: a single shared client in src/lib/posthog.ts
(initPostHog + exported posthog) initialized inside a 'use client'
PostHogProvider mounted in the root layout. init and all captures now share one
instance. $pageview is captured manually on initial load + every route change
(usePathname only, to keep the ~800 static pages static). track() routes its
dual-write through the same module.

Verified locally: with init in the provider, the browser actually POSTs events
— /_ph/e/, /_ph/i/v0/e/ (pageview), and /_ph/s/ (recording) all return 200.
(Confirmed by temporarily disabling the bot-UA filter so the headless test
browser wasn't filtered; that flag was reverted before commit.)

Removes instrumentation-client.ts and components/posthog-pageview.tsx.
